requirements.txt
# Streamlit Cloud compatible requirements with fixed versions
streamlit==1.29.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
PyYAML>=5.3
requests>=2.23.0
scipy>=1.4.1
pandas>=1.1.4
matplotlib>=3.2.2
seaborn>=0.11.0
thop>=0.1.0
tqdm>=4.64.0
streamlit-webrtc>=0.45.0
av>=9.2.0